Updated hit-box and textures
I've created textures for the player in an 8-bit design including a new texture for the bullet. 
I adjusted the position of the bullet to visually appear closer to the center of the airship.

The png sizes took up too much space so i changed the sizes. 
I've added a call to have a border drawn on the player's hit-box and edited the player's hit-box closer to the outline of the airship.